% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_favefood.R
\name{age_favefood}
\alias{age_favefood}
\title{Age & Favorite Food}
\usage{
age_favefood(name, age, fave_food)
}
\arguments{
\item{name}{a character containing the input name}

\item{age}{a number containing a 1 - 3 digit age}

\item{fave_food}{a character containing the input favorite food}
}
\value{

}
\description{
Contains one function age_favefood that prints some information about the user. It requests as input, the user's name, age and their favorite food.
}
