% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beach_weather.R
\name{beach_weather}
\alias{beach_weather}
\title{Beach Weather}
\usage{
beach_weather(temp, cloudcover)
}
\arguments{
\item{temp}{accepts a digit that represents the temperature in Fahrenheit}

\item{cloudcover}{accepts a digit that represents the cloud cover in Oktas}
}
\value{

}
\description{
This function takes temperature in Fahrenheit and cloud cover in oktas as inputs and returns a message on whether or not it is a good beach day
}
