% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/need_caffeine.R
\name{need_caffeine}
\alias{need_caffeine}
\title{Need Caffeine}
\usage{
need_caffeine(fatigue_level, work_remaining)
}
\arguments{
\item{fatigue_level}{Takes a double as input on a scale of 1 - 10 to represent your level of fatigue that day}

\item{work_remaining}{Takes a double as input on a scale of 1 - 10 to represent your amount of work remaining to be completed that day}
}
\value{

}
\description{
This function takes as input your level of fatigue and the amount of work remaining in the day and returns whether or not you should drink some coffee.
}
