% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grayblue.R
\name{theme_grayblue}
\alias{theme_grayblue}
\title{GrayBlue Theme}
\usage{
theme_grayblue()
}
\value{

}
\description{
This function was named for the colors of the lines on the ggplot, with the major lines being gray and the minor lines being blue.
}
